src/requirements.txt
click==8.1.7
fastapi==0.115.0
uvicorn==0.31.0
docker==7.1.0
pyyaml==6.0.2
httpx==0.27.2